Some changes

+++ Application/Handlers/DeleteHandlers/DeleteStudentHandler.cs
@@ -10,6 +10,8 @@ namespace Application.Handlers.DeleteHandlers
     {
         public async Task<Student> Handle(DeleteStudentDto dto)
         {
+            if(!await DbContext.Set<Student>().AnyAsync(f => f.ID == dto.ID))
+                throw new Exception("Student not found");
             await DbContext.Set<Student>().Where(f => f.ID == dto.ID)
                 .ExecuteDeleteAsync();
             return default;
